+++ src/main/java/smartup/microservices/dto/DemandeInfoPersoDto.java
@@ -1,8 +1,5 @@
 package smartup.microservices.dto;
 
-import java.util.Date;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 import lombok.AllArgsConstructor;
@@ -17,7 +14,7 @@ import lombok.Setter;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class DemandeInfoPersoDto {
 	
-	private int employerInfosPersoRhDemLibelle;
+	private Long employerInfosPersoRhDemLibelle;
 	private String employerInfosPersoRhDemMatricule;
 	private String employerInfosPersoRhDemUsername;
 	private String employerInfosPersoRhDemAdress;
