+++ src/main/java/smartup/microservices/services/DemandeFraisProfesService.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 
 import smartup.microservices.entities.DemandeFraisProfes;
-import smartup.microservices.entities.DemandeFraisProfes;
 
 public interface DemandeFraisProfesService {
 	public DemandeFraisProfes addDemandeFraisProfes(DemandeFraisProfes dat);
@@ -12,7 +11,7 @@ public interface DemandeFraisProfesService {
 	public DemandeFraisProfes updateDemandeFraisProfes(DemandeFraisProfes dat);
 	public Optional<DemandeFraisProfes> retrieveDemandeFraisProfes(Long id);
 	public DemandeFraisProfes getDemandeFraisProfesById(Long id);
-	public void deleteDemandeFraisProfesById(int id);
+	public void deleteDemandeFraisProfesById(Long id);
 	void acceptedemandeById(Long id);
 	DemandeFraisProfes activerDemandeFraisProfes(Long id);
 	DemandeFraisProfes desactiverDemandeFraisProfes(Long id);
